+++ src/app/pages/new-poem/step-3/WeightsPage.ts
@@ -4,7 +4,7 @@ import { Router } from "@angular/router"
 // Import Local Components
 import { WeightOption } from "./WeightOption"
 import { Steps } from "../steps/steps"
-import { ReturnButton } from "../return-back/ReturnButton"
+import { ReturnButton } from "../shared/ReturnButton"
 
 // Import Local Services
 import { Forms } from "../../../services/forms.service"
